loading and saving
the websocket will no attempt to retrieve  device data from local file "Devices.json" before launching and write any changes directly to the file. failure to locate file will result in the server crashing. packet forwarding hasnt been fully implemented yet and for the time being will retrieve test data instead. this is not yet send back to the website as that part is depending in the implementation of packet forwarding.