Web Interface Data Request Update
Web Interface is now able to request either all devices, all data of a specific device or any specific data set of a specific device. Buttons SHOULD enable and disable as the respective features become available or unavable but this feature hasnt been fully tested yet. Buttons are on page load initiated to there default states. Web Interface will automatically in chain request first identification followed by a request of all devices upon connection. The javascript version of the websocket has been updated alongside while the C# rewrite is still in process. The Websocket will now only assign a new identification if the IP hasnt been identified in the current session before and thus eliminate any duplicate issues. The websocket will for now serve dummy data as the work on the devices has just begun. Removed old Prototype device scripts as they are all getting major reworks, also removed currently unused Files "minecraft.json", "computercraft.json" and "todo.txt". The first two because they are currently not used until the 3D rendering gets implemented and the last one because its a dublicate of both my local "To-do.md" abd the Wiki´s To-do List. Currently there is no error handling during the phase of identification, if a connecting Web Interface or Device doesnt follow the Protocol exactly it will crash the Websocket. A Fix to this will be included with either the 3D Rendering Update or the Instruction set/chain update later after the device scripts have catched up to the progression of the web interface and the websocket.